diffuse ibl

+++ Project 2/Project 2/hdr_fbo.hpp
@@ -21,7 +21,7 @@ struct HDRFBO
 
 			glGenTextures(1, &colorBuffer);
 			glBindTexture(GL_TEXTURE_2D, colorBuffer);
-			glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, nullptr);
+			glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, nullptr);
 			glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 			glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
